refactor(components/atom/label): migrate to ts
