+++ components/atom/label/src/settings.tsx
@@ -1,0 +1,16 @@
+export const CLASSNAME = 'sui-AtomLabel'
+
+export const TYPES = {
+  SUCCESS: 'success',
+  ERROR: 'error',
+  ALERT: 'alert',
+  CONTRAST: 'contrast',
+  DISABLED: 'disabled'
+} as const
+
+export const FONT_SIZES = {
+  LARGE: 'large',
+  MEDIUM: 'medium',
+  SMALL: 'small',
+  XSMALL: 'xsmall'
+} as const
