+++ components/atom/label/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../../tsconfig.json",
+  "compilerOptions": {
+    "types": ["react"],
+    "outDir": "./lib",
+    "rootDir": "./src"
+  },
+  "exclude": ["lib"]
+}
